requirements.txt
httpx[http2]
pandas
jinja2
python-dotenv
